fix: resolve 6 miscellaneous lint warnings
- play_manager.c: uppercase literal suffix (100.0f → 100.0F)
- audio_processor_beep.c: isolate declarations (split 6 variables to separate lines)
- play_manager.c: non-const parameter (made src_block const in resample_audio_block)
- bt_manager.c: signed-char-misuse (cast rssi via unsigned char)
- audio_util.c: narrowing conversion (explicit cast for channels_in)
- beep_manager.h: inconsistent parameter name (cb → callback to match implementation)

Total: 6 warnings fixed
Build: SUCCESS
Progress: 241/761 warnings (31.7%)

+++ esp_bt_audio_source/components/audio_processor/audio_util.c
@@ -118,7 +118,7 @@ esp_err_t resample_audio(const audio_resample_args_t *args)
     }
 
     int channels = (channels_in == AUDIO_CHANNEL_MONO || channels_in == AUDIO_CHANNEL_STEREO)
-                       ? channels_in
+                       ? (int)channels_in
                        : AUDIO_CHANNEL_STEREO;
 
     const int bps = bytes_per_sample(bit_depth);
